@@ -19,7 +19,7 @@
     "autoprefixer": "10.4.14",
     "eslint": "^8.42.0",
     "eslint-config-next": "^13.4.5",
-    "next": "^14.2.3",
+    "next": "^15.5.9",
     "postcss": "^8.4.24",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
