@@ -18,4 +18,4 @@ const nextConfig = {
   },
 }
 
-module.exports = nextConfig
+export default nextConfig
